backend/task_management/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
python-decouple>=3.8
channels>=4.1.0
daphne>=4.1.2
psycopg2-binary>=2.9

